backend/src/requirements.txt
Django==4.2.1
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-import-export==3.2.0
django-postgres-extra==2.0.8
django-split-settings==1.2.0
django_cors_headers==4.3.0
djangorestframework==3.14.0
drf-yasg==1.21.5
psycopg==3.1.9  # apt install libpq5
python-dotenv==1.0.0
pytz==2023.3
jsonschema==4.19.0
PyJWT==2.8.0
pydantic[dotenv,email]==1.10.9
cryptography==41.0.3
backoff==2.2.1
